+++ src/main/java/kitchenpos/menu/application/MenuGroupService.java
@@ -1,12 +1,12 @@
-package kitchenpos.application;
+package kitchenpos.menu.application;
 
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.domain.MenuGroup;
-import kitchenpos.dto.response.MenuGroupResponse;
+import kitchenpos.menu.domain.MenuGroup;
+import kitchenpos.menu.dto.response.MenuGroupResponse;
 import kitchenpos.exception.ErrorCode;
 import kitchenpos.exception.KitchenposException;
-import kitchenpos.repository.MenuGroupRepository;
+import kitchenpos.menu.domain.MenuGroupRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
