+++ src/main/java/kitchenpos/menu/domain/MenuRepository.java
@@ -1,7 +1,7 @@
-package kitchenpos.repository;
+package kitchenpos.menu.domain;
 
 import java.util.List;
-import kitchenpos.domain.Menu;
+import kitchenpos.menu.domain.Menu;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MenuRepository extends JpaRepository<Menu, Long> {
