+++ src/test/java/kitchenpos/menu/domain/MenuProductTest.java
@@ -1,8 +1,9 @@
-package kitchenpos.domain;
+package kitchenpos.menu.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
+import kitchenpos.menu.domain.MenuProduct;
 import org.junit.jupiter.api.Test;
 
 class MenuProductTest {
