+++ src/test/java/kitchenpos/menu/ui/MenuRestControllerTest.java
@@ -1,6 +1,6 @@
-package kitchenpos.ui;
+package kitchenpos.menu.ui;
 
-import static kitchenpos.ui.MenuGroupRestControllerTest.메뉴_그룹_생성_요청;
+import static kitchenpos.menu.ui.MenuGroupRestControllerTest.메뉴_그룹_생성_요청;
 import static kitchenpos.ui.ProductRestControllerTest.상품_생성_요청;
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -8,11 +8,11 @@ import java.math.BigDecimal;
 import java.util.Arrays;
 import java.util.List;
 import kitchenpos.BaseTest;
-import kitchenpos.domain.MenuGroup;
-import kitchenpos.dto.request.MenuProductRequest;
-import kitchenpos.dto.request.MenuRequest;
+import kitchenpos.menu.domain.MenuGroup;
+import kitchenpos.menu.dto.request.MenuProductRequest;
+import kitchenpos.menu.dto.request.MenuRequest;
 import kitchenpos.dto.request.ProductRequest;
-import kitchenpos.dto.response.MenuResponse;
+import kitchenpos.menu.dto.response.MenuResponse;
 import kitchenpos.dto.response.ProductResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -21,7 +21,7 @@ import org.springframework.http.HttpMethod;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 
-class MenuRestControllerTest extends BaseTest {
+public class MenuRestControllerTest extends BaseTest {
     private final ProductRequest 상품 = new ProductRequest("후라이드", BigDecimal.valueOf(16000));
     private MenuProductRequest 메뉴_항목;
     private final MenuGroup 메뉴_그룹 = new MenuGroup("한마리메뉴");
